API

SpectralSuperpositions.AbstractSuperposition — Type

Abstract type for creating 2-D spectra structs. The types and arguments have been defined as well as a validate_superposition function to ensure consistency.

SpectralSuperpositions.AbstractSuperposition1D — Type

Abstract type for creating 2-D spectra structs. The types and arguments have been defined as well as a validate_superposition function to ensure consistency.

SpectralSuperpositions.axesinfo — Function

axesinfo()
axesinfo(s::Symbol)
axesinfo(x)

Query the axis type and its physical dimensions.

There are 8 supported spectral-variable types, formed from the combinations of spatial/temporal domain, linear/angular geometry, and frequency/period quantity. There's one direction type :direction, for a total of 9 possible axis types.

axesinfo() returns a dictionary with information for all 9 possible axis types.

axesinfo(s::Symbol) returns information for a specific axis type, e.g. axesinfo(:wavenumber)

axesinfo(x) return the axis information for the axes of a spectrum x.

The dictionary is as follows:

:direction          => ((:direction,), 𝐀),
:frequency          => ((:temporal, :linear, :frequency), 𝐓^-1),
:angular_frequency  => ((:temporal, :angular, :frequency), 𝐀 * 𝐓^-1),
:period             => ((:temporal, :linear, :period), 𝐓),
:angular_period     => ((:temporal, :angular, :period), 𝐓 * 𝐀^-1),
:wavenumber         => ((:spatial, :linear, :frequency), 𝐋^-1),
:angular_wavenumber => ((:spatial, :angular, :frequency), 𝐀 * 𝐋^-1),
:wavelength         => ((:spatial, :linear, :period), 𝐋),
:angular_wavelength => ((:spatial, :angular, :period), 𝐋 * 𝐀^-1)

SpectralSuperpositions.axesnames — Function

axesnames(x::AbstractSuperposition)
axesnames(x::AbstractSuperposition1D)

Return the axes names for an AbstractSuperposition or the axis name for an AbstractSuperposition1D

SpectralSuperpositions.axestypes — Function

axestypes(x::Quantity)
axestypes(x::Units)
axestypes(x::AbstractVector{<:Quantity})
axestypes(x::AbstractSuperposition)
axestypes(x::AbstractSuperposition1D)
axestypes(dim::Dimensions)

Returns the type of axes as a symbol based on the unitful dimension. Uses the following dictionary.

Dict{Unitful.Dimensions, Symbol} with 9 entries:
𝐋⁻¹   => :wavenumber
𝐀 𝐋⁻¹ => :angular_wavenumber
𝐓⁻¹   => :frequency
𝐀     => :direction
𝐀 𝐓⁻¹ => :angular_frequency
𝐋     => :wavelength
𝐓     => :period
𝐓 𝐀⁻¹ => :angular_period
𝐋 𝐀⁻¹ => :angular_wavelength

SpectralSuperpositions.coordinates — Method

coordinates(x::AbstractSuperposition)

Returns a symbol whether the Spectrum is in :polar or :cartesian coordinates.

SpectralSuperpositions.evenspacing — Function

evenspacing(x::AbstractVector)
evenspacing(x::AbstractSuperposition)
evenspacing(x::AbstractSuperposition1D)

Return evenly spaced axis parameters as tuples (start, step, length).

For a Spectrum two such tuples are returned, one for each axis. Throws ArgumentError if the axis or axes are not evenly spaced.

SpectralSuperpositions.isangular — Method

isangular(x)

Returns true if the unit dimensions are :angular.

SpectralSuperpositions.iscartesian — Method

iscartesian(x::AbstractSuperposition)

Returns true if the Spectrum is in :cartesian coordinates.

SpectralSuperpositions.isdirection — Method

isdirection(x)

Returns true if the unit dimensions are a :direction.

SpectralSuperpositions.isevenlyspaced — Function

isevenlyspaced(x::AbstractVector)
isevenlyspaced(x::AbstractSuperposition)
isevenlyspaced(x::AbstractSuperposition1D)

Return true if the axis values are evenly spaced.

SpectralSuperpositions.isfrequency — Method

isfrequency(x)

Returns true if the unit dimensions are :frequency.

SpectralSuperpositions.islinear — Method

islinear(x)

Returns true if the unit dimensions are :linear.

SpectralSuperpositions.isperiod — Method

isperiod(x)

Returns true if the unit dimensions are :period.

SpectralSuperpositions.ispolar — Method

ispolar(x::AbstractSuperposition)

Returns true if the Spectrum is in :polar coordinates.

SpectralSuperpositions.isspatial — Method

isspatial(x::AbstractSuperposition)

Returns true if the first axis of AbstractSuperposition has unit dimensions that are :spatial.

SpectralSuperpositions.isspatial — Method

isspatial(x)

Returns true if the unit dimensions are :spatial.

SpectralSuperpositions.isspectralvariable — Method

isspectralvariable(x)

Returns true if the unit dimensions are not a direction and some other spectral variable instead.

SpectralSuperpositions.istemporal — Method

istemporal(x::AbstractSuperposition)

Returns true if the first axis of AbstractSuperposition has unit dimensions that are :temporal.

SpectralSuperpositions.istemporal — Method

istemporal(x)

Returns true if the unit dimensions are :temporal.

See also axesinfo()

SpectralSuperpositions.superposition_unit_aliases — Method

superposition_unit_aliases(::AbstractSuperposition1D)

Default symbol for getting the units of the data.

SpectralSuperpositions.superposition_unit_aliases — Method

superposition_unit_aliases(::AbstractSuperposition)

Default symbol for getting the units of the data.

SpectralSuperpositions.validate_superposition — Method

validate_superposition(
    data::AbstractMatrix,
    axis1::AbstractVector{<:Quantity},
    axis2::AbstractVector{<:Quantity}
)

Recommended validation function for structs that inherit from AbstractSuperposition. Ensures type consistency, axis/data sizes, axes values are increasing, moves spectral variables to first axis, and saves useful attributes like coordinates, axestypes, and axesnames.

SpectralSuperpositions.validate_superposition1d — Method

validate_superposition1d(
    data::AbstractVector,
    axis::AbstractVector{<:Quantity}
)

Recommended validation function for structs that inherit from AbstractSuperposition1D. Ensures type consistency, axis/data sizes, axes values are increasing, moves spectral variables to first axis, and saves useful attributes like coordinates, axestypes, and axesnames.

Unitful.unit — Method

unit(x::AbstractSuperposition, quantity::Symbol)
unit(x::AbstractSuperposition)

Extend Unitful.unit for abstract superposition.

The quantity can be :axis (or the axis name) or :integral. These return the units of the spectral-variable axis, integral quantity, or the superposition data quantity. The default is quantity=:superposition.

Unitful.unit — Method

unit(x::AbstractSuperposition1D, quantity::Symbol)
unit(x::AbstractSuperposition1D)

Extend Unitful.unit for abstract superposition.

The quantity can be :axis (or the axis name) or :integral. These return the units of the spectral-variable axis, integral quantity, or the superposition data quantity. The default is quantity=:superposition.